requirements.txt
click==6.7
Flask==1.0.2
Flask-Cors==3.0.4
Flask-Uploads==0.2.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.3
Pillow==5.1.0
six==1.11.0
Werkzeug==0.14.1
